Release INAV 2.1

+++ src/main/common/filter.h
@@ -55,13 +55,13 @@ typedef float (*filterApplyFnPtr)(void *filter, float input);
 
 float nullFilterApply(void *filter, float input);
 
-void pt1FilterInit(pt1Filter_t *filter, uint8_t f_cut, float dT);
+void pt1FilterInit(pt1Filter_t *filter, float f_cut, float dT);
 void pt1FilterInitRC(pt1Filter_t *filter, float tau, float dT);
 void pt1FilterSetTimeConstant(pt1Filter_t *filter, float tau);
 float pt1FilterGetLastOutput(pt1Filter_t *filter);
 float pt1FilterApply(pt1Filter_t *filter, float input);
 float pt1FilterApply3(pt1Filter_t *filter, float input, float dT);
-float pt1FilterApply4(pt1Filter_t *filter, float input, uint16_t f_cut, float dt);
+float pt1FilterApply4(pt1Filter_t *filter, float input, float f_cut, float dt);
 void pt1FilterReset(pt1Filter_t *filter, float input);
 
 void rateLimitFilterInit(rateLimitFilter_t *filter);
